requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
pillow>=10.0.0
requests>=2.31.0
sentencepiece>=0.1.99
protobuf>=3.20.0